refactor: dedup token-meta + approval-chain boilerplate
Three behavior-preserving extractions across the prepare_* action builders.

- Hoist the multicall(decimals, symbol) pattern from three local copies
  (compound/actions, morpho/actions, execution/index) into a single
  shared/token-meta.ts helper.
- Hoist the if (approval) { walk tail; attach; return } tail-attach
  pattern that every approve-then-act builder hand-rolled (9 sites across
  compound, morpho, positions, staking) into shared/approval.ts as
  chainApproval().
- Inline the parseAmountFriendly wrapper in positions/actions (it just
  called parseUnits); drop its dead re-export.
- Remove the getLedgerStatus wrapper that only forwarded to
  getSessionStatus; alias the import at the registration site instead.

Net -75 lines across 7 files (+52 / -149). No tool, schema, or response
shape changed. vitest: 442/442 pass.
