+++ src/modules/compound/actions.ts
@@ -1,8 +1,8 @@
 import { encodeFunctionData, parseUnits, maxUint256 } from "viem";
 import { cometAbi } from "../../abis/compound-comet.js";
-import { erc20Abi } from "../../abis/erc20.js";
 import { getClient } from "../../data/rpc.js";
-import { buildApprovalTx, resolveApprovalCap } from "../shared/approval.js";
+import { buildApprovalTx, chainApproval, resolveApprovalCap } from "../shared/approval.js";
+import { resolveTokenMeta } from "../shared/token-meta.js";
 import type {
   PrepareCompoundSupplyArgs,
   PrepareCompoundWithdrawArgs,
@@ -10,21 +10,6 @@ import type {
   PrepareCompoundRepayArgs,
 } from "./schemas.js";
 import type { SupportedChain, UnsignedTx } from "../../types/index.js";
-
-async function resolveMeta(
-  chain: SupportedChain,
-  asset: `0x${string}`
-): Promise<{ decimals: number; symbol: string }> {
-  const client = getClient(chain);
-  const [decimals, symbol] = await client.multicall({
-    contracts: [
-      { address: asset, abi: erc20Abi, functionName: "decimals" },
-      { address: asset, abi: erc20Abi, functionName: "symbol" },
-    ],
-    allowFailure: false,
-  });
-  return { decimals: Number(decimals), symbol: symbol as string };
-}
 
 async function resolveBaseToken(
   chain: SupportedChain,
@@ -76,7 +61,7 @@ export async function buildCompoundSupply(p: PrepareCompoundSupplyArgs): Promise
   const asset = p.asset as `0x${string}`;
   const wallet = p.wallet as `0x${string}`;
   await assertCometActionAllowed(chain, market, "supply");
-  const meta = await resolveMeta(chain, asset);
+  const meta = await resolveTokenMeta(chain, asset);
   const amountWei = parseUnits(p.amount, meta.decimals);
   const { approvalAmount, display } = resolveApprovalCap(
     p.approvalCap,
@@ -107,13 +92,7 @@ export async function buildCompoundSupply(p: PrepareCompoundSupplyArgs): Promise
     description: `Supply ${p.amount} ${meta.symbol} to Compound V3 ${market} on ${chain}`,
     decoded: { functionName: "supply", args: { asset, amount: p.amount, market } },
   };
-  if (approval) {
-    let tail = approval;
-    while (tail.next) tail = tail.next;
-    tail.next = supplyTx;
-    return approval;
-  }
-  return supplyTx;
+  return chainApproval(approval, supplyTx);
 }
 
 export async function buildCompoundWithdraw(p: PrepareCompoundWithdrawArgs): Promise<UnsignedTx> {
@@ -122,7 +101,7 @@ export async function buildCompoundWithdraw(p: PrepareCompoundWithdrawArgs): Pro
   const asset = p.asset as `0x${string}`;
   const wallet = p.wallet as `0x${string}`;
   await assertCometActionAllowed(chain, market, "withdraw");
-  const meta = await resolveMeta(chain, asset);
+  const meta = await resolveTokenMeta(chain, asset);
   const amountWei = p.amount === "max" ? maxUint256 : parseUnits(p.amount, meta.decimals);
   return {
     chain,
@@ -146,7 +125,7 @@ export async function buildCompoundBorrow(p: PrepareCompoundBorrowArgs): Promise
   const wallet = p.wallet as `0x${string}`;
   await assertCometActionAllowed(chain, market, "withdraw");
   const baseToken = await resolveBaseToken(chain, market);
-  const meta = await resolveMeta(chain, baseToken);
+  const meta = await resolveTokenMeta(chain, baseToken);
   const amountWei = parseUnits(p.amount, meta.decimals);
   return {
     chain,
@@ -170,7 +149,7 @@ export async function buildCompoundRepay(p: PrepareCompoundRepayArgs): Promise<U
   const wallet = p.wallet as `0x${string}`;
   await assertCometActionAllowed(chain, market, "supply");
   const baseToken = await resolveBaseToken(chain, market);
-  const meta = await resolveMeta(chain, baseToken);
+  const meta = await resolveTokenMeta(chain, baseToken);
   const amountWei = p.amount === "max" ? maxUint256 : parseUnits(p.amount, meta.decimals);
   let approval: UnsignedTx | null = null;
   if (amountWei !== maxUint256) {
@@ -204,11 +183,5 @@ export async function buildCompoundRepay(p: PrepareCompoundRepayArgs): Promise<U
     description: `Repay ${p.amount === "max" ? "all" : p.amount} ${meta.symbol} on Compound V3 ${market} on ${chain}`,
     decoded: { functionName: "supply(base)", args: { asset: baseToken, amount: p.amount, market } },
   };
-  if (approval) {
-    let tail = approval;
-    while (tail.next) tail = tail.next;
-    tail.next = repayTx;
-    return approval;
-  }
-  return repayTx;
+  return chainApproval(approval, repayTx);
 }
